Add wording to simplify icon size selection
In the previous version, I didn't take into account that there are icons that can be categorized as both square and non-square.

In the new version, I added a brief instruction so that contributors will always know what size to choose for an icon. It will also make the review process a bit easier, since it is easier to estimate the density of an icon than the squareness.