Add UI for icon requesting
This commits adds a simple UI to view all unthemed icons, alongside a way to either copy them to clipboard and open the form (either prefilled or empty).

It also adds a simple long-click gesture to directly visit the icon request form with prefilled information.

If the URL length is greater than DIRECT_LINK_MAX_LENGTH however, it removes the direct link and offers the user to copy the contents to the clipboard.

Fixes #2162